Create Dicegame.java
One of your friends' uncle who owns a Casino is struck with a problem. He wants to quickly calculate the amount a player would receive while he finishes the game. You would want to help him as well as get couple of free chances in the Casino. Show your expertise.
The rules are as follows:

A player can enter the game with minimum amount of 1000.
A player can take any number of chances. If the value of the dice is '1', the player leaves the game with half the amount. If the value is other than '1', the total amount is raised by multiple of 100*value of dice.
If the player quits, he/she leaves the game with amount he/she has got.