Create Paritygenerator.java
Parity is a very useful tool in information processing in digital computers to indicate any presence of error in bit information. To indicate any occurrence of error, an extra bit is included with the message based on the total number of 1s in a set of data, which is called parity. The parity bit is generated based on the number of 1's in the message.

For even parity the parity bit generated is '1' when the number of ones in the message is odd and '0' when the number of ones in the message is even.

And for odd parity the parity bit generated is '0' when the number of ones in the message is odd and '1' when the number of ones in the message is even.

 Find the parity bit generated for a 8 bit message.