Create Thresholdorceil.java
One of the most common representations of a two dimensional array is an Image. Most of the images are represented in the form of matrix and algorithms are applied on the matrix / 2-D Array.  Given an image with the pixel values, manipulate the image by changing the values to the corresponding threshold values. 

Input contains image size and three threshold values.
Change all the values that are less than threshold value 1 to the corresponding threshold value.
Change all the values that are in between threshold 1 and threshold 2 to the threshold value 2.
Change all the values that are greater than threshold value 2 to the threshold value 3.